Shifting button from purchases list to purchase status
