+++ frontend/manage/app/purchase/purchase-status.component.html
@@ -5,17 +5,40 @@
     <li class="breadcrumb-item"><a routerLink="/purchases">Purchases</a></li>
     <li class="breadcrumb-item active">Purchase {{ purchase.id }}</li>
   </ol>
-
+  <h2>{{purchase.publication.title}}</h2>
+  <div *ngIf="purchase.status == 'ok'">
+      <button (click)="onDownload_LCPL(purchase)"
+        type="button" role="button"
+        class="btn btn-secondary btn-sm">
+        <i class="fa fa-download" aria-hidden="true"></i> LCPL
+      </button>
+      <a *ngIf="purchase.type == 'LOAN' && purchase.licenseUuid != null"
+        [routerLink]="['/purchases', purchase.id, 'renew']"
+        class="btn btn-secondary btn-sm" role="button">
+        <i class="fa fa-hourglass-end" aria-hidden="true"></i> Renew
+      </a>
+      <button *ngIf="purchase.type == 'LOAN' && purchase.licenseUuid != null"
+          type="button" role="button"
+          class="btn btn-secondary btn-sm"
+          (click)="onReturn(purchase)">
+        <i class="fa fa-eject" aria-hidden="true"></i> Return
+      </button>
+      <p>Date : {{formatDate(purchase.transactionDate)}}</p>
+      <p>User : {{purchase.user.name}}</p>
+      <p>Type : {{purchase.type}}</p>
+      <p *ngIf="purchase.type == 'LOAN'">Start Date : {{formatDate(purchase.startDate)}}</p>
+      <p *ngIf="purchase.type == 'LOAN'">End Date : {{formatDate(purchase.endDate)}}</p>
+      <p>Delivred : {{purchase.licenseUuid != null}}<p>
   <div *ngIf="licenseStatus">
     <form>
       <fieldset disabled>
-        <div class="form-group">
+        <div *ngIf="licenseStatus" class="form-group">
           <label>Status:</label>
           <input type="text" class="form-control" placeholder="{{ licenseStatus.status }}" />
         </div>
         <div *ngIf="licenseStatus.potential_rights" class="form-group">
           <label>Potential Right End Date:</label>
-          <input type="text" class="form-control" placeholder="{{ licenseStatus.potential_rights.end }}" />
+          <input  type="text" class="form-control" placeholder="{{ licenseStatus.potential_rights.end }}" />
         </div>
       </fieldset>
     </form>
